address review feedback: fix jina tier, extract OpenAI deep base class
- Fix jina-search tier from ai-grounded to raw-search and remove from
  comprehensive group (Greptile P0)
- Extract OpenAIDeepBaseProvider to eliminate duplication between
  openai-deep and openai-deep-o3 adapters (Greptile P2)
- Fix biome formatting issues (import ordering, line length)
